requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
pypdf
langchain
langchain-community
langchain-text-splitters
openai
tiktoken
openai 
tiktoken 
langchain 
langchain-community
langchain-openai
